Fixups
Mostly changing ints to NSInteger and fixing up deprecated constants

+++ Managed Objects/TSSTManagedGroup.m
@@ -286,18 +286,18 @@
 	if(!solidDirectory)
 	{
 		[groupLock lock];
-		imageData = [[self instance] contentsOfEntry: index];
+		imageData = [[self instance] contentsOfEntry: (int) index];
 		[groupLock unlock];
 	}
 	else
 	{
-		NSString * name = [[self instance] nameOfEntry: index];
+		NSString * name = [[self instance] nameOfEntry: (int) index];
 		NSString * fileName = [NSString stringWithFormat:@"%li.%@", (long)index, [name pathExtension]];
 		fileName = [solidDirectory stringByAppendingPathComponent: fileName];
 		if(![[NSFileManager defaultManager] fileExistsAtPath: fileName])
 		{
 			[groupLock lock];
-			imageData = [[self instance] contentsOfEntry: index];
+			imageData = [[self instance] contentsOfEntry: (int) index];
 			[groupLock unlock];
 			[imageData writeToFile: fileName options: 0 error: nil];
 		}
@@ -513,7 +513,7 @@
     NSPDFImageRep * rep = [self instance];
     TSSTPage * imageDescription;
     NSMutableSet * pageSet = [NSMutableSet set];
-    int imageCount = [rep pageCount];
+    NSInteger imageCount = [rep pageCount];
     int pageNumber;
     for (pageNumber = 0; pageNumber < imageCount; ++pageNumber)
     {
